feat: auto suggestion locations dev complete

+++ app/src/main/java/com/example/daggerhilttest/remote/WeatherApi.kt
@@ -22,15 +22,15 @@ interface WeatherApi {
 
     @GET("weather")
     suspend fun getCurrentWeatherByLatLong(
-        @Query("lat") lat: Float,
-        @Query("lon") long: Float,
+        @Query("lat") lat: Double,
+        @Query("lon") long: Double,
         @Query("appid") apiKey: String
     ): Response<CurrentWeather>
 
     @GET("forecast")
     suspend fun getHourlyForecastByLatLong(
-        @Query("lat") lat: Float,
-        @Query("lon") long: Float,
+        @Query("lat") lat: Double,
+        @Query("lon") long: Double,
         @Query("appid") apiKey: String
     ): Response<HourlyForecast>
 }
